+++ zoomimage-compose/src/commonMain/kotlin/com/github/panpf/zoomimage/compose/subsampling/internal/ComposeSubsamplingUtils.kt
@@ -16,11 +16,6 @@
 
 package com.github.panpf.zoomimage.compose.subsampling.internal
 
-import androidx.compose.runtime.Composable
-import com.github.panpf.zoomimage.subsampling.StoppedController
 import com.github.panpf.zoomimage.subsampling.internal.TileBitmapConvertor
-
-@Composable
-expect fun defaultStoppedController(): StoppedController?
 
 expect fun createTileBitmapConvertor(): TileBitmapConvertor?